Moved Column settings to class ColumnSettings
